update: milestone! hello world stdout!

+++ generator/arm64/assembler_arm64.h
@@ -21,7 +21,7 @@ enum SectionType {
 struct StackVar {
     const char *varName;
     int varSize;
-    int stackOffset;
+    uint64_t stackOffset;
 };
 
 /**
